requirements.txt
python-dotenv
apscheduler
aiohttp
asyncio